pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragshield"
version = "0.1.0"
description = "Prompt injection detection for RAG pipelines using ZEDD (Zero-Shot Embedding Drift Detection)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "RagShield Contributors" }
]
keywords = [
    "llm",
    "rag",
    "prompt-injection",
    "security",
    "embeddings",
    "langchain",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
    "typing_extensions>=4.0.0;python_version<'3.10'",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
langchain = ["langchain-core>=0.1.0"]
llamaindex = ["llama-index-core>=0.10.0"]
cli = ["typer>=0.9.0", "rich>=13.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = ["ragshield[openai,langchain,cli]"]

[project.scripts]
ragshield = "ragshield.cli.main:app"

[project.urls]
Homepage = "https://github.com/ragshield/ragshield"
Documentation = "https://github.com/ragshield/ragshield#readme"
Repository = "https://github.com/ragshield/ragshield"
Issues = "https://github.com/ragshield/ragshield/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ragshield"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ragshield --cov-report=term-missing"
